pyproject.toml
[project]
name = "iam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cachetools>=7.0.1",
    "cryptography>=46.0.4",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=24.1.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
